redoxr

Module redoxr

Redoxr is a open-source build_scripter inspired by nob.h and Cargo.

Its not made to be used in actual projects and does not fit many use cases. On the contrary it does help with actually thinking about what you are building and what libs you really need. There is also (planned) cross-compatitibility with Cargo both ways. Cargo to Redox will be added first and Redox to Cargo will be added using Truck (see at the bottom). All outputs will be put in the bin/ (and bin/deps/) directory, which sanitizes the environment and adds reproducability.

This is just the start and in the end I want to be able to compile anything using Redoxr and also be able to easily extend it to any other language using “intelligent trait usage”.

Basic Usage:

mod redoxr;
use redoxr::redoxr::*;

fn main() -> () {
    let mut redoxr = Redoxr::new();

    let mut main_crate = RustCrate::main( "some_crate");
    if let Some(error) = main_crate.compile() {panic!("{}",error)};
}

You can add dependencies just by using .depend_on(dependency). (For this you currently need to compile the dependency first):

mod redoxr;
use redoxr::redoxr::*;

fn main() -> () {
    let mut redoxr = Redoxr::new();
    let mut dependency = RustCrate::from_cargo("clap", "clap");

    if let Some(error) = dependency.compile() {panic!("{}",error)}

    let mut main_crate = RustCrate::main(&mut redoxr, "some_crate").
        .depend_on(dependecy.clone());

    //There is also this macro that compiles dependencies.
    compile!(main_crate);
}

The compile! macro just expands to the whole if-let-then-panic-statement.

The dependency list uses the incredible Mirror technology (just a semi-safe wrapper around raw pointer), which allows for dependencies to be built after being passed as such while still being able to handle errors. Currently that is most of the magic, but soon I will implement the version that uses traits, making it expandable to other languages easily as long as I write it correctly.

Structs

Mirror

Redoxr

RustCrate

A Struct that defines a Rust Crate managed by any build system

Enums

RedoxError

Constants

COMP_VERSION

PATH_SEPERATOR

Traits

RedoxrCompatible

Type Aliases

AnyError

Cmd

MainResult